submissions/available/NNSlicer/requirements.txt
absl-py==0.7.1
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.1
astropy==2.0.3
atari-py==0.2.6
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
cleverhans==3.0.1
click==6.7
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==2.6.0
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
faiss==1.5.3
fastcache==1.0.2
filelock==2.0.13
fire==0.2.1
Flask==0.12.2
Flask-Cors==3.0.3
flatbuffers==1.11
foolbox==1.8.0
funcsigs==1.0.2
future==0.18.1
gast==0.2.2
genpac==2.1.0
gevent==1.2.2
gitdb2==3.0.2
GitPython==3.0.8
glob2==0.6
gmpy2==2.0.8
google-pasta==0.1.7
greenlet==0.4.12
grpcio==1.22.0
gym==0.15.3
h5py==2.7.1
heapdict==1.0.0
horovod==0.18.1
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imagesize==0.7.1
inflection==0.3.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mnist==0.2.2
mock==3.0.5
mpmath==1.0.0
msgpack==1.0.0
msgpack-numpy==0.4.4.3
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
netaddr==0.7.19
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpy==1.17.2
numpydoc==0.7.0
odo==0.5.1
olefile==0.46
opencv-python==4.1.0.25
opencv-python-headless==4.1.1.26
openpyxl==2.4.10
opt-einsum==3.1.0
packaging==16.8
pandas==0.24.2
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.10.0
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
py-spy==0.2.2
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.1
pydotplus==2.0.2
pyflakes==1.6.0
pyglet==1.3.2
Pygments==2.2.0
pylint==1.7.2
Pympler==0.8
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
randomgen==1.14
ray==0.7.2
redis==3.3.11
requests==2.18.4
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.20.0
scipy==1.0.0
seaborn==0.8.1
selenium==3.141.0
Send2Trash==1.4.2
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.5
tblib==1.3.2
tensorboard==1.13.1
tensorboardX==1.8
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.8.0
tensorpack==0.10
termcolor==1.1.0
terminado==0.8.1
terminaltables==3.1.0
testpath==0.3.1
text-unidecode==1.3
toolz==0.9.0
torch==1.1.0
torchvision==0.3.0
tornado==4.5.3
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.11.2
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
